fix: redesign webhook reconcile/stop lifecycle boundary
Collapse stacked admission gates into a single life context cancelled by
Stop, keep a simple coalescing deferred reconciler, and move lifecycle
contracts into dedicated suites under the 300-line growth limit.

Generated-By: looper 0.11.4 (runner=fixer, agent=grok-build)
