+++ internal/runtime/runtime_test.go
@@ -7,7 +7,6 @@ import (
 	"errors"
 	"net"
 	"os"
-	"os/exec"
 	"path/filepath"
 	"strconv"
 	"strings"
@@ -235,371 +234,6 @@ func TestRuntimeProjectMutationsAtomicallyPublishCatalog(t *testing.T) {
 	}
 	if afterPublishCalls != 2 {
 		t.Fatalf("project follow-up calls after remove = %d, want 2", afterPublishCalls)
-	}
-}
-
-func TestRuntimeProjectAddDoesNotWaitForWebhookReconciliation(t *testing.T) {
-	t.Parallel()
-
-	workingDir := t.TempDir()
-	cfg, err := config.DefaultConfig(workingDir)
-	if err != nil {
-		t.Fatalf("DefaultConfig() error = %v", err)
-	}
-	cfg.Storage.DBPath = filepath.Join(workingDir, "runtime.sqlite")
-	cfg.Webhook.Enabled = true
-	cfg.Webhook.Mode = config.WebhookModeTunnel
-	cfg.Webhook.ListenPort = 0
-	cfg.Webhook.PublicBaseURL = "https://looper.example.test"
-	ghPath := "/usr/bin/gh"
-	cfg.Tools.GHPath = &ghPath
-
-	rt := New(Options{
-		Config:           cfg,
-		Logger:           &testLogger{},
-		RunSchedulerTick: func(context.Context, Services) error { return nil },
-	})
-	if err := rt.Start(context.Background()); err != nil {
-		t.Fatalf("Start() error = %v", err)
-	}
-	t.Cleanup(func() { rt.Stop("test cleanup") })
-
-	reconcileStarted := make(chan struct{})
-	releaseReconcile := make(chan struct{})
-	defer close(releaseReconcile)
-	rt.webhook.bootstrapDone = true
-	rt.webhook.tunnelClient = blockingWebhookTunnelGitHubClient{
-		started: reconcileStarted,
-		release: releaseReconcile,
-	}
-
-	projectService := rt.Services().Projects
-	projectService.ListWorktrees = nil
-	repo := "acme/live"
-	addDone := make(chan error, 1)
-	go func() {
-		_, err := projectService.AddProject(context.Background(), projects.AddInput{
-			ID: "live", Name: "Live", RepoPath: workingDir, Repo: &repo, SnapshotMode: projects.SnapshotModeOff,
-		})
-		addDone <- err
-	}()
-
-	select {
-	case <-reconcileStarted:
-	case <-time.After(time.Second):
-		t.Fatal("webhook reconciliation did not start")
-	}
-	select {
-	case err := <-addDone:
-		if err != nil {
-			t.Fatalf("AddProject() error = %v", err)
-		}
-	case <-time.After(time.Second):
-		t.Fatal("AddProject() waited for webhook reconciliation")
-	}
-}
-
-// Contract: while tunnel reconcile from mutation A is blocked on GitHub, mutation B
-// may publish the catalog (updateConfig) without racing the in-flight pass.
-// Each Reconcile uses the config snapshot captured at pass start; the coalesced
-// follow-up pass reconciles the second project after the first unblocks.
-func TestRuntimeSecondProjectPublishDuringBlockedTunnelReconcile(t *testing.T) {
-	t.Parallel()
-
-	workingDir := t.TempDir()
-	secondDir := t.TempDir()
-	cfg, err := config.DefaultConfig(workingDir)
-	if err != nil {
-		t.Fatalf("DefaultConfig() error = %v", err)
-	}
-	cfg.Storage.DBPath = filepath.Join(workingDir, "runtime.sqlite")
-	cfg.Webhook.Enabled = true
-	cfg.Webhook.Mode = config.WebhookModeTunnel
-	cfg.Webhook.ListenPort = 0
-	cfg.Webhook.PublicBaseURL = "https://looper.example.test"
-	ghPath := "/usr/bin/gh"
-	cfg.Tools.GHPath = &ghPath
-
-	rt := New(Options{
-		Config:           cfg,
-		Logger:           &testLogger{},
-		RunSchedulerTick: func(context.Context, Services) error { return nil },
-	})
-	if err := rt.Start(context.Background()); err != nil {
-		t.Fatalf("Start() error = %v", err)
-	}
-	t.Cleanup(func() { rt.Stop("test cleanup") })
-
-	firstStarted := make(chan struct{})
-	releaseFirst := make(chan struct{})
-	secondCreateSeen := make(chan struct{})
-	client := &recordingBlockingTunnelClient{
-		firstStarted:     firstStarted,
-		firstRelease:     releaseFirst,
-		secondCreateSeen: secondCreateSeen,
-	}
-	rt.webhook.bootstrapDone = true
-	rt.webhook.tunnelClient = client
-
-	projectService := rt.Services().Projects
-	projectService.ListWorktrees = nil
-	repoA := "acme/first"
-	if _, err := projectService.AddProject(context.Background(), projects.AddInput{
-		ID: "first", Name: "First", RepoPath: workingDir, Repo: &repoA, SnapshotMode: projects.SnapshotModeOff,
-	}); err != nil {
-		t.Fatalf("AddProject(first) error = %v", err)
-	}
-
-	select {
-	case <-firstStarted:
-	case <-time.After(2 * time.Second):
-		t.Fatal("first tunnel reconcile CreateHook did not start")
-	}
-
-	// First pass must still be blocked on GitHub when the second mutation publishes.
-	if got := client.createRepos(); len(got) != 1 || got[0] != repoA {
-		t.Fatalf("creates while first blocked = %v, want only %q", got, repoA)
-	}
-
-	repoB := "acme/second"
-	addSecondDone := make(chan error, 1)
-	go func() {
-		_, err := projectService.AddProject(context.Background(), projects.AddInput{
-			ID: "second", Name: "Second", RepoPath: secondDir, Repo: &repoB, SnapshotMode: projects.SnapshotModeOff,
-		})
-		addSecondDone <- err
-	}()
-
-	select {
-	case err := <-addSecondDone:
-		if err != nil {
-			t.Fatalf("AddProject(second) error = %v", err)
-		}
-	case <-time.After(2 * time.Second):
-		t.Fatal("AddProject(second) waited for blocked first tunnel reconciliation")
-	}
-
-	// Concurrent publish must not advance CreateHook for B until first pass ends.
-	if got := client.createRepos(); len(got) != 1 || got[0] != repoA {
-		t.Fatalf("creates after second publish while first blocked = %v, want only %q", got, repoA)
-	}
-
-	close(releaseFirst)
-
-	select {
-	case <-secondCreateSeen:
-	case <-time.After(3 * time.Second):
-		t.Fatal("coalesced reconcile did not CreateHook for second project")
-	}
-
-	got := client.createRepos()
-	if len(got) < 2 || got[0] != repoA {
-		t.Fatalf("create order = %v, want first %q then later %q", got, repoA, repoB)
-	}
-	foundB := false
-	for _, repo := range got[1:] {
-		if repo == repoB {
-			foundB = true
-			break
-		}
-	}
-	if !foundB {
-		t.Fatalf("create order = %v, want %q after first pass unblocked", got, repoB)
-	}
-}
-
-// Contract: project add schedules deferred webhook reconcile; Stop must cancel
-// that pass (and not hang on wg) even while the GitHub tunnel boundary blocks.
-func TestRuntimeStopCancelsBlockedScheduledWebhookReconcile(t *testing.T) {
-	t.Parallel()
-
-	workingDir := t.TempDir()
-	cfg, err := config.DefaultConfig(workingDir)
-	if err != nil {
-		t.Fatalf("DefaultConfig() error = %v", err)
-	}
-	cfg.Storage.DBPath = filepath.Join(workingDir, "runtime.sqlite")
-	cfg.Webhook.Enabled = true
-	cfg.Webhook.Mode = config.WebhookModeTunnel
-	cfg.Webhook.ListenPort = 0
-	cfg.Webhook.PublicBaseURL = "https://looper.example.test"
-	ghPath := "/usr/bin/gh"
-	cfg.Tools.GHPath = &ghPath
-
-	rt := New(Options{
-		Config:           cfg,
-		Logger:           &testLogger{},
-		RunSchedulerTick: func(context.Context, Services) error { return nil },
-	})
-	if err := rt.Start(context.Background()); err != nil {
-		t.Fatalf("Start() error = %v", err)
-	}
-
-	reconcileStarted := make(chan struct{})
-	// Intentionally never release: Stop must cancel via stop-aware context.
-	releaseReconcile := make(chan struct{})
-	rt.webhook.bootstrapDone = true
-	rt.webhook.tunnelClient = blockingWebhookTunnelGitHubClient{
-		started: reconcileStarted,
-		release: releaseReconcile,
-	}
-
-	projectService := rt.Services().Projects
-	projectService.ListWorktrees = nil
-	repo := "acme/live"
-	if _, err := projectService.AddProject(context.Background(), projects.AddInput{
-		ID: "live", Name: "Live", RepoPath: workingDir, Repo: &repo, SnapshotMode: projects.SnapshotModeOff,
-	}); err != nil {
-		t.Fatalf("AddProject() error = %v", err)
-	}
-
-	select {
-	case <-reconcileStarted:
-	case <-time.After(time.Second):
-		t.Fatal("webhook reconciliation did not start")
-	}
-
-	stopDone := make(chan struct{})
-	go func() {
-		rt.Stop("test blocked reconcile stop")
-		close(stopDone)
-	}()
-
-	select {
-	case <-stopDone:
-	case <-time.After(2 * time.Second):
-		t.Fatal("Stop() blocked on scheduled webhook reconciliation")
-	}
-}
-
-// Contract: project mutation → admitted gh-forward launch (past execCommand /
-// pipe setup) → Stop must still refuse cmd.Start when shutdown wins the
-// start-admission boundary, and must not hang on the in-flight launch.
-func TestRuntimeStopRejectsScheduledGHForwardLaunch(t *testing.T) {
-	// Mutates package-level execCommand / pre-start / started hooks.
-	workingDir := t.TempDir()
-	cfg, err := config.DefaultConfig(workingDir)
-	if err != nil {
-		t.Fatalf("DefaultConfig() error = %v", err)
-	}
-	cfg.Storage.DBPath = filepath.Join(workingDir, "runtime.sqlite")
-	cfg.Webhook.Enabled = true
-	cfg.Webhook.Mode = config.WebhookModeGHForward
-	ghPath := "/usr/bin/gh"
-	cfg.Tools.GHPath = &ghPath
-
-	testBin, err := os.Executable()
-	if err != nil {
-		t.Fatalf("os.Executable() error = %v", err)
-	}
-
-	var (
-		startGate   = make(chan struct{})
-		launchSeen  = make(chan struct{})
-		startedHook = make(chan struct{}, 1)
-	)
-	originalCommand := execCommand
-	originalStartedHook := webhookForwarderStartedHook
-	originalPreStartHook := webhookForwarderPreStartHook
-	execCommand = func(name string, args ...string) *exec.Cmd {
-		// Command construction alone must not be the stop boundary; admission
-		// is exercised after pipes are attached, inside admitForwarderStart.
-		cmd := exec.Command(testBin, "-test.run=TestWebhookRuntimeForwarderHelperProcess", "--")
-		cmd.Env = append(os.Environ(), "GO_WANT_HELPER_PROCESS=1")
-		cmd.Args[0] = name
-		return cmd
-	}
-	webhookForwarderPreStartHook = func() {
-		select {
-		case <-launchSeen:
-		default:
-			close(launchSeen)
-		}
-		<-startGate
-	}
-	webhookForwarderStartedHook = func() {
-		select {
-		case startedHook <- struct{}{}:
-		default:
-		}
-	}
-	t.Cleanup(func() {
-		execCommand = originalCommand
-		webhookForwarderStartedHook = originalStartedHook
-		webhookForwarderPreStartHook = originalPreStartHook
-		select {
-		case <-startGate:
-		default:
-			close(startGate)
-		}
-	})
-
-	rt := New(Options{
-		Config:           cfg,
-		Logger:           &testLogger{},
-		RunSchedulerTick: func(context.Context, Services) error { return nil },
-	})
-	if err := rt.Start(context.Background()); err != nil {
-		t.Fatalf("Start() error = %v", err)
-	}
-	rt.webhook.bootstrapDone = true
-
-	projectService := rt.Services().Projects
-	projectService.ListWorktrees = nil
-	repo := "acme/gh-forward-live"
-	if _, err := projectService.AddProject(context.Background(), projects.AddInput{
-		ID: "ghfwd", Name: "GH Forward Live", RepoPath: workingDir, Repo: &repo, SnapshotMode: projects.SnapshotModeOff,
-	}); err != nil {
-		t.Fatalf("AddProject() error = %v", err)
-	}
-
-	select {
-	case <-launchSeen:
-	case <-time.After(2 * time.Second):
-		// If reconcile never reached launch (catalog/config path), still exercise
-		// Stop after an explicit ScheduleReconcile with the published catalog.
-		rt.scheduleWebhookForwarderReconcile()
-		select {
-		case <-launchSeen:
-		case <-time.After(2 * time.Second):
-			t.Fatal("scheduled gh-forward reconcile did not attempt forwarder launch")
-		}
-	}
-
-	stopDone := make(chan struct{})
-	go func() {
-		rt.Stop("test gh-forward launch stop")
-		close(stopDone)
-	}()
-
-	// Wait until the webhook runtime has observed shutdown before releasing the
-	// start gate. Runtime.Stop drains config/scheduler/etc. before
-	// stopWebhookRuntime; a fixed sleep races on slow CI and can let
-	// admitForwarderStart win the mutex and start the process.
-	select {
-	case <-rt.webhook.stopCh:
-	case <-time.After(3 * time.Second):
-		t.Fatal("webhook runtime did not observe Stop before start gate release")
-	}
-	close(startGate)
-
-	select {
-	case <-stopDone:
-	case <-time.After(3 * time.Second):
-		t.Fatal("Stop() blocked on scheduled gh-forward forwarder launch")
-	}
-
-	select {
-	case <-startedHook:
-		t.Fatal("forwarder process reached started hook after Stop began")
-	default:
-	}
-
-	status := rt.WebhookStatus()
-	for _, fwd := range status.Forwarders {
-		if fwd.Running || fwd.PID != nil {
-			t.Fatalf("forwarder still running after Stop: %#v", fwd)
-		}
 	}
 }
 
@@ -5178,117 +4812,4 @@ func waitForCondition(t *testing.T, timeout time.Duration, condition func() bool
 		time.Sleep(25 * time.Millisecond)
 	}
 	t.Fatal("condition not satisfied before timeout")
-}
-
-type blockingWebhookTunnelGitHubClient struct {
-	started chan<- struct{}
-	release <-chan struct{}
-}
-
-func (c blockingWebhookTunnelGitHubClient) GetHook(context.Context, string, int64) (webhookTunnelGitHubHook, bool, error) {
-	return webhookTunnelGitHubHook{}, false, nil
-}
-
-func (c blockingWebhookTunnelGitHubClient) CreateHook(ctx context.Context, _ string, _ string, _ string, _ []string) (webhookTunnelGitHubHook, error) {
-	close(c.started)
-	select {
-	case <-c.release:
-		return webhookTunnelGitHubHook{ID: 1}, nil
-	case <-ctx.Done():
-		return webhookTunnelGitHubHook{}, ctx.Err()
-	}
-}
-
-func (c blockingWebhookTunnelGitHubClient) UpdateHook(context.Context, string, int64, string, string, []string, bool) (webhookTunnelGitHubHook, error) {
-	return webhookTunnelGitHubHook{}, nil
-}
-
-func (c blockingWebhookTunnelGitHubClient) DeleteHook(context.Context, string, int64) error {
-	return nil
-}
-
-// recordingBlockingTunnelClient blocks only the first CreateHook, remembers
-// created hooks for later GetHook, and signals when a second create arrives.
-type recordingBlockingTunnelClient struct {
-	mu               sync.Mutex
-	hooks            map[string]webhookTunnelGitHubHook
-	createOrder      []string
-	nextID           int64
-	firstStarted     chan struct{}
-	firstRelease     <-chan struct{}
-	secondCreateSeen chan struct{}
-	firstSignaled    bool
-}
-
-func (c *recordingBlockingTunnelClient) createRepos() []string {
-	c.mu.Lock()
-	defer c.mu.Unlock()
-	out := make([]string, len(c.createOrder))
-	copy(out, c.createOrder)
-	return out
-}
-
-func (c *recordingBlockingTunnelClient) GetHook(_ context.Context, repo string, id int64) (webhookTunnelGitHubHook, bool, error) {
-	c.mu.Lock()
-	defer c.mu.Unlock()
-	hook, ok := c.hooks[repo]
-	if !ok || hook.ID != id {
-		return webhookTunnelGitHubHook{}, false, nil
-	}
-	return hook, true, nil
-}
-
-func (c *recordingBlockingTunnelClient) CreateHook(ctx context.Context, repo string, url string, _ string, events []string) (webhookTunnelGitHubHook, error) {
-	c.mu.Lock()
-	c.nextID++
-	id := c.nextID
-	hook := webhookTunnelGitHubHook{ID: id, Active: true, Events: append([]string(nil), events...)}
-	hook.Config.URL = url
-	hook.Config.ContentType = "json"
-	hook.Config.InsecureSSL = "0"
-	if c.hooks == nil {
-		c.hooks = map[string]webhookTunnelGitHubHook{}
-	}
-	c.hooks[repo] = hook
-	c.createOrder = append(c.createOrder, repo)
-	createIndex := len(c.createOrder)
-	first := !c.firstSignaled
-	if first {
-		c.firstSignaled = true
-	}
-	c.mu.Unlock()
-
-	if first {
-		close(c.firstStarted)
-		select {
-		case <-c.firstRelease:
-		case <-ctx.Done():
-			return webhookTunnelGitHubHook{}, ctx.Err()
-		}
-	} else if createIndex == 2 {
-		select {
-		case <-c.secondCreateSeen:
-		default:
-			close(c.secondCreateSeen)
-		}
-	}
-	return hook, nil
-}
-
-func (c *recordingBlockingTunnelClient) UpdateHook(_ context.Context, repo string, id int64, url string, _ string, events []string, active bool) (webhookTunnelGitHubHook, error) {
-	c.mu.Lock()
-	defer c.mu.Unlock()
-	hook := webhookTunnelGitHubHook{ID: id, Active: active, Events: append([]string(nil), events...)}
-	hook.Config.URL = url
-	hook.Config.ContentType = "json"
-	hook.Config.InsecureSSL = "0"
-	if c.hooks == nil {
-		c.hooks = map[string]webhookTunnelGitHubHook{}
-	}
-	c.hooks[repo] = hook
-	return hook, nil
-}
-
-func (c *recordingBlockingTunnelClient) DeleteHook(context.Context, string, int64) error {
-	return nil
 }
