+++ internal/runtime/webhook_runtime_test.go
@@ -9,7 +9,6 @@ import (
 	"os/exec"
 	"strings"
 	"sync"
-	"sync/atomic"
 	"testing"
 	"time"
 
@@ -401,302 +400,6 @@ func TestWebhookRuntimeReconcilePassesDeadlineToTunnelHookReconcile(t *testing.T
 	}
 }
 
-func TestWebhookRuntimeEnsureTunnelServerRefusesAfterStop(t *testing.T) {
-	t.Parallel()
-
-	_, _, cfg := setupWebhookTunnelTestRepos(t)
-	cfg.Webhook.ListenPort = 0
-	rt := newWebhookRuntime(cfg, &testLogger{}, time.Now)
-	rt.Stop()
-
-	if err := rt.ensureTunnelServer(); !errors.Is(err, errWebhookRuntimeStopped) {
-		t.Fatalf("ensureTunnelServer() after Stop error = %v, want %v", err, errWebhookRuntimeStopped)
-	}
-	if rt.tunnelServer != nil {
-		t.Fatal("tunnelServer started after Stop, want nil")
-	}
-}
-
-func TestWebhookRuntimeLaunchForwarderRefusesAfterStop(t *testing.T) {
-	// Mutates package-level execCommand; must not run in parallel.
-	var starts atomic.Int32
-	originalCommand := execCommand
-	execCommand = func(name string, args ...string) *exec.Cmd {
-		starts.Add(1)
-		return exec.Command("false")
-	}
-	t.Cleanup(func() { execCommand = originalCommand })
-
-	rt := &webhookRuntime{
-		status: WebhookStatus{
-			Enabled:    true,
-			Forwarders: []WebhookForwarderState{{Repo: "nexu-io/looper", Command: []string{"gh", "webhook", "forward"}}},
-		},
-		stopCh:          make(chan struct{}),
-		forwarderStopCh: map[string]chan struct{}{"nexu-io/looper": make(chan struct{})},
-		now:             time.Now,
-	}
-	rt.Stop()
-	rt.launchForwarder("nexu-io/looper")
-
-	// launchForwarder must not admit after Stop: no command construction and no
-	// extra work that would race with Stop's completed Wait.
-	time.Sleep(50 * time.Millisecond)
-	if got := starts.Load(); got != 0 {
-		t.Fatalf("execCommand starts after Stop = %d, want 0", got)
-	}
-}
-
-// Contract: Stop that wins the pre-start boundary must refuse cmd.Start under
-// the same lock that observes stopped; no process side effect after shutdown.
-func TestWebhookRuntimeAdmitForwarderStartRefusesAfterStop(t *testing.T) {
-	testBin, err := os.Executable()
-	if err != nil {
-		t.Fatalf("os.Executable() error = %v", err)
-	}
-
-	startGate := make(chan struct{})
-	atBoundary := make(chan struct{})
-	originalPreStart := webhookForwarderPreStartHook
-	webhookForwarderPreStartHook = func() {
-		close(atBoundary)
-		<-startGate
-	}
-	t.Cleanup(func() {
-		webhookForwarderPreStartHook = originalPreStart
-		select {
-		case <-startGate:
-		default:
-			close(startGate)
-		}
-	})
-
-	rt := &webhookRuntime{
-		stopCh: make(chan struct{}),
-		now:    time.Now,
-	}
-	cmd := exec.Command(testBin, "-test.run=TestWebhookRuntimeForwarderHelperProcess", "--")
-	cmd.Env = append(os.Environ(), "GO_WANT_HELPER_PROCESS=1")
-
-	errCh := make(chan error, 1)
-	go func() {
-		errCh <- rt.admitForwarderStart(cmd)
-	}()
-
-	select {
-	case <-atBoundary:
-	case <-time.After(2 * time.Second):
-		t.Fatal("admitForwarderStart did not reach pre-start boundary")
-	}
-
-	rt.Stop()
-	close(startGate)
-
-	select {
-	case err := <-errCh:
-		if !errors.Is(err, errWebhookRuntimeStopped) {
-			t.Fatalf("admitForwarderStart() error = %v, want errWebhookRuntimeStopped", err)
-		}
-	case <-time.After(2 * time.Second):
-		t.Fatal("admitForwarderStart did not return after Stop")
-	}
-	if cmd.Process != nil {
-		t.Fatal("cmd.Process started after Stop won start admission")
-	}
-}
-
-// Contract: Stop that wins the pre-persist boundary must refuse store.Upsert
-// before any database write begins.
-func TestWebhookRuntimeAdmitForwarderPersistRefusesAfterStop(t *testing.T) {
-	repositories := openWebhookRuntimeTestRepositories(t)
-	persistGate := make(chan struct{})
-	atBoundary := make(chan struct{})
-	originalPrePersist := webhookForwarderPrePersistHook
-	webhookForwarderPrePersistHook = func() {
-		close(atBoundary)
-		<-persistGate
-	}
-	t.Cleanup(func() {
-		webhookForwarderPrePersistHook = originalPrePersist
-		select {
-		case <-persistGate:
-		default:
-			close(persistGate)
-		}
-	})
-
-	rt := &webhookRuntime{
-		stopCh:         make(chan struct{}),
-		now:            time.Now,
-		forwarderStore: repositories.WebhookForwarders,
-	}
-	record := storage.WebhookForwarderRecord{
-		Repo: "nexu-io/looper", PID: 1, ProcessStart: 1, Fingerprint: "fp",
-		Endpoint: "http://127.0.0.1/webhook/forward", Events: "push", GHPath: "/usr/bin/gh",
-		DaemonID: "daemon", SpawnedAt: 1, UpdatedAt: 1,
-	}
-
-	errCh := make(chan error, 1)
-	go func() {
-		errCh <- rt.admitForwarderPersist(repositories.WebhookForwarders, record)
-	}()
-
-	select {
-	case <-atBoundary:
-	case <-time.After(2 * time.Second):
-		t.Fatal("admitForwarderPersist did not reach pre-persist boundary")
-	}
-
-	rt.Stop()
-	close(persistGate)
-
-	select {
-	case err := <-errCh:
-		if !errors.Is(err, errWebhookRuntimeStopped) {
-			t.Fatalf("admitForwarderPersist() error = %v, want errWebhookRuntimeStopped", err)
-		}
-	case <-time.After(2 * time.Second):
-		t.Fatal("admitForwarderPersist did not return after Stop")
-	}
-
-	rows, err := repositories.WebhookForwarders.List(context.Background())
-	if err != nil {
-		t.Fatalf("WebhookForwarders.List() error = %v", err)
-	}
-	if len(rows) != 0 {
-		t.Fatalf("persisted %d forwarder rows after Stop, want 0", len(rows))
-	}
-}
-
-// Contract: after persist admission succeeds, Stop must still set stopped and
-// close stopCh without waiting for the (possibly blocked) database write.
-// Completing the write after Stop must not leave a durable forwarder row.
-func TestWebhookRuntimeStopUnblockedDuringAdmittedPersist(t *testing.T) {
-	repositories := openWebhookRuntimeTestRepositories(t)
-	persistGate := make(chan struct{})
-	atAdmitted := make(chan struct{})
-	originalAdmitted := webhookForwarderAdmittedPersistHook
-	webhookForwarderAdmittedPersistHook = func() {
-		close(atAdmitted)
-		<-persistGate
-	}
-	t.Cleanup(func() {
-		webhookForwarderAdmittedPersistHook = originalAdmitted
-		select {
-		case <-persistGate:
-		default:
-			close(persistGate)
-		}
-	})
-
-	rt := &webhookRuntime{
-		stopCh:         make(chan struct{}),
-		now:            time.Now,
-		forwarderStore: repositories.WebhookForwarders,
-	}
-	record := storage.WebhookForwarderRecord{
-		Repo: "nexu-io/looper", PID: 1, ProcessStart: 1, Fingerprint: "fp",
-		Endpoint: "http://127.0.0.1/webhook/forward", Events: "push", GHPath: "/usr/bin/gh",
-		DaemonID: "daemon", SpawnedAt: 1, UpdatedAt: 1,
-	}
-
-	errCh := make(chan error, 1)
-	go func() {
-		errCh <- rt.admitForwarderPersist(repositories.WebhookForwarders, record)
-	}()
-
-	select {
-	case <-atAdmitted:
-	case <-time.After(2 * time.Second):
-		t.Fatal("admitForwarderPersist did not reach post-admission persist boundary")
-	}
-
-	stopDone := make(chan struct{})
-	go func() {
-		rt.Stop()
-		close(stopDone)
-	}()
-
-	select {
-	case <-stopDone:
-	case <-time.After(2 * time.Second):
-		t.Fatal("Stop blocked while admitted forwarder persist held only the DB path")
-	}
-	select {
-	case <-rt.stopCh:
-	default:
-		t.Fatal("stopCh not closed after Stop during admitted persist")
-	}
-
-	close(persistGate)
-
-	select {
-	case err := <-errCh:
-		if !errors.Is(err, errWebhookRuntimeStopped) {
-			t.Fatalf("admitForwarderPersist() error = %v, want errWebhookRuntimeStopped after Stop", err)
-		}
-	case <-time.After(2 * time.Second):
-		t.Fatal("admitForwarderPersist did not return after Stop released admitted gate")
-	}
-
-	rows, err := repositories.WebhookForwarders.List(context.Background())
-	if err != nil {
-		t.Fatalf("WebhookForwarders.List() error = %v", err)
-	}
-	if len(rows) != 0 {
-		t.Fatalf("persisted %d forwarder rows after admitted write followed by Stop, want 0", len(rows))
-	}
-}
-
-func TestWebhookRuntimeReconcileDoesNotLaunchForwarderAfterStop(t *testing.T) {
-	// Mutates package-level execCommand; must not run in parallel.
-	repositories := openWebhookRuntimeTestRepositories(t)
-	nowISO := formatJavaScriptISOString(time.Date(2026, time.May, 16, 12, 0, 0, 0, time.UTC))
-	metadata := `{"repo":"nexu-io/looper"}`
-	if err := repositories.Projects.Upsert(context.Background(), storage.ProjectRecord{ID: "project_1", Name: "Looper", RepoPath: "/tmp/looper", MetadataJSON: &metadata, CreatedAt: nowISO, UpdatedAt: nowISO}); err != nil {
-		t.Fatalf("Projects.Upsert() error = %v", err)
-	}
-
-	var starts atomic.Int32
-	originalCommand := execCommand
-	execCommand = func(name string, args ...string) *exec.Cmd {
-		starts.Add(1)
-		return exec.Command("false")
-	}
-	t.Cleanup(func() { execCommand = originalCommand })
-
-	rt := &webhookRuntime{
-		cfg:    webhookRuntimeTestConfig("nexu-io/looper"),
-		ghPath: "/usr/bin/gh",
-		status: WebhookStatus{
-			Enabled:     true,
-			EndpointURL: "http://127.0.0.1:7777/webhook/forward",
-		},
-		stopCh:          make(chan struct{}),
-		forwarderStopCh: map[string]chan struct{}{},
-		now:             time.Now,
-		bootstrapDone:   true,
-	}
-	rt.Stop()
-
-	if err := rt.Reconcile(repositories); err != nil {
-		t.Fatalf("Reconcile() after Stop error = %v", err)
-	}
-	time.Sleep(50 * time.Millisecond)
-	if got := starts.Load(); got != 0 {
-		t.Fatalf("execCommand starts after stopped Reconcile = %d, want 0", got)
-	}
-	if status := rt.Status(); len(status.Forwarders) != 0 {
-		// Reconcile may still materialize desired state before launch admission;
-		// either empty or non-running is fine as long as no process started.
-		for _, fwd := range status.Forwarders {
-			if fwd.Running || fwd.PID != nil {
-				t.Fatalf("forwarder started after Stop: %#v", fwd)
-			}
-		}
-	}
-}
-
 func TestWebhookRuntimeReconcileClearsTransientListFailureAfterRecovery(t *testing.T) {
 	testBin, err := os.Executable()
 	if err != nil {
@@ -917,100 +620,6 @@ func TestWebhookRuntimeReconcileUsesCapturedCatalogSnapshotWhenForgejoPublishes(
 
 // Contract: tunnel helpers must keep using the Reconcile-captured config while a
 // concurrent project publication rewrites w.cfg.Projects under updateConfig.
-func TestWebhookRuntimeTunnelReconcileUsesCapturedConfigAcrossUpdateConfig(t *testing.T) {
-	t.Parallel()
-
-	_, repos, cfg := setupWebhookTunnelTestRepos(t)
-	cfg.Webhook.Mode = config.WebhookModeTunnel
-	cfg.Webhook.PublicBaseURL = "https://captured.example.test"
-	cfg.Projects = []config.ProjectRefConfig{{ID: "first", Repo: "acme/first"}}
-	rt := newWebhookRuntime(cfg, &testLogger{}, func() time.Time { return time.Unix(10, 0) })
-	rt.ghPath = "/usr/bin/gh"
-	rt.bootstrapDone = true
-	t.Cleanup(rt.Stop)
-
-	started := make(chan struct{})
-	release := make(chan struct{})
-	var (
-		createdRepo atomic.Value
-		createdURL  atomic.Value
-	)
-	rt.tunnelClient = &callbackTunnelCreateClient{
-		onCreate: func(_ context.Context, repo, url string) (webhookTunnelGitHubHook, error) {
-			createdRepo.Store(repo)
-			createdURL.Store(url)
-			close(started)
-			<-release
-			return webhookTunnelGitHubHook{ID: 11, Active: true, Events: webhookForwardEvents}, nil
-		},
-	}
-
-	captured := rt.configSnapshot()
-	done := make(chan error, 1)
-	go func() {
-		done <- rt.reconcileSnapshot(context.Background(), repos, captured)
-	}()
-
-	select {
-	case <-started:
-	case <-time.After(2 * time.Second):
-		t.Fatal("tunnel CreateHook did not start")
-	}
-
-	// Publish a second project while the first pass is blocked inside tunnel helpers.
-	// updateConfig mutates w.cfg.Projects under w.mu; helpers must not reread w.cfg.
-	published := captured
-	published.Projects = append(append([]config.ProjectRefConfig(nil), captured.Projects...), config.ProjectRefConfig{ID: "second", Repo: "acme/second"})
-	for i := 0; i < 100; i++ {
-		rt.updateConfig(published)
-	}
-
-	close(release)
-	select {
-	case err := <-done:
-		if err != nil {
-			t.Fatalf("reconcileSnapshot() error = %v", err)
-		}
-	case <-time.After(2 * time.Second):
-		t.Fatal("reconcileSnapshot did not finish after release")
-	}
-
-	if got, _ := createdRepo.Load().(string); got != "acme/first" {
-		t.Fatalf("CreateHook repo = %q, want acme/first from captured snapshot", got)
-	}
-	wantURL := webhookTunnelManagedURL(captured, "acme/first")
-	if got, _ := createdURL.Load().(string); got != wantURL {
-		t.Fatalf("CreateHook URL = %q, want captured snapshot URL %q", got, wantURL)
-	}
-	status := rt.Status()
-	if len(status.TunnelHooks) != 1 || status.TunnelHooks[0].Repo != "acme/first" {
-		t.Fatalf("status.TunnelHooks = %#v, want only captured repo acme/first", status.TunnelHooks)
-	}
-}
-
-type callbackTunnelCreateClient struct {
-	onCreate func(ctx context.Context, repo, url string) (webhookTunnelGitHubHook, error)
-}
-
-func (c *callbackTunnelCreateClient) GetHook(context.Context, string, int64) (webhookTunnelGitHubHook, bool, error) {
-	return webhookTunnelGitHubHook{}, false, nil
-}
-
-func (c *callbackTunnelCreateClient) CreateHook(ctx context.Context, repo, url, _ string, events []string) (webhookTunnelGitHubHook, error) {
-	if c.onCreate != nil {
-		return c.onCreate(ctx, repo, url)
-	}
-	return webhookTunnelGitHubHook{ID: 1, Active: true, Events: events}, nil
-}
-
-func (c *callbackTunnelCreateClient) UpdateHook(context.Context, string, int64, string, string, []string, bool) (webhookTunnelGitHubHook, error) {
-	return webhookTunnelGitHubHook{}, nil
-}
-
-func (c *callbackTunnelCreateClient) DeleteHook(context.Context, string, int64) error {
-	return nil
-}
-
 func TestWebhookRuntimeTerminalForwarderExitLatchesWithoutRespawn(t *testing.T) {
 	testBin, err := os.Executable()
 	if err != nil {
